% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secondp.R
\name{coef.secondp}
\alias{coef.secondp}
\title{\code{coef.secondp}: Extract coefficients from the model fit}
\usage{
\method{coef}{secondp}(object, ...)
}
\arguments{
\item{object}{An \code{secondp} object}

\item{...}{Other \code{coef} arguments}
}
\value{
Coefficients in the OLS model
}
\description{
S3 method \code{coef} for an S3 object of class \code{secondp}
}
