% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gib_chain_rcpp}
\alias{gib_chain_rcpp}
\title{A Gibbs sampler using Rcpp}
\usage{
gib_chain_rcpp(a, b, m)
}
\arguments{
\item{a}{parameter of beta distribution}

\item{b}{parameter of beta distribution}

\item{m}{parameter of distribution}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
