% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secondp.R
\name{print.secondp}
\alias{print.secondp}
\title{\code{print.secondp}: Print variable selection results}
\usage{
\method{print}{secondp}(x, ...)
}
\arguments{
\item{x}{An \code{secondp} object}

\item{...}{Other \code{print} arguments}
}
\value{
Variable selection results
}
\description{
S3 method \code{print} for an S3 object of class \code{secondp}
}
