% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secondp.R
\name{secondp}
\alias{secondp}
\title{\code{Secondp} function}
\usage{
secondp(x, y)
}
\arguments{
\item{x}{Independent variables, can be a \code{matrix} or a \code{data.frame}}

\item{y}{Dependent variable, can be a \code{vector} or a column from a \code{data.frame}}
}
\value{
A list of following components:
\describe{
\item{var.index}{A vector of indices of selected variables}
\item{var.label}{A vector of labels of selected variables}
\item{x}{Input data \code{x}}
\item{y}{Input data \code{y}}
\item{null.bound}{Null bound in the SGPV screening}
\item{pe.can}{Point estimates in the candidate set}
\item{lb.can}{Lower bounds of CI in the candidate set}
\item{ub.can}{Upper bounds of CI in the candidate set}
}
}
\description{
This function outputs the variable selection results using Second-Generation P-values.
}
\seealso{
\itemize{
\item \code{\link[=print.secondp]{print.secondp()}} prints the variable selection results
\item \code{\link[=coef.secondp]{coef.secondp()}} extracts coefficient estimates
\item \code{\link[=summary.secondp]{summary.secondp()}} summarizes the OLS outputs
}
}
