\name{StatComp21036-package}
\alias{StatComp21036-package}
\alias{StatComp21036}
\docType{package}
\title{\packageTitle{StatComp21036}}
\description{\packageDescription{StatComp21036}}
\section{Package Content}{\packageIndices{StatComp21036}}
\author{\packageAuthor{StatComp21036}}
\section{Maintainer}{\packageMaintainer{StatComp21036}}
\keyword{package}
