% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secondp.R
\name{summary.secondp}
\alias{summary.secondp}
\title{\code{summary.secondp}: Summary of the final model}
\usage{
\method{summary}{secondp}(object, ...)
}
\arguments{
\item{object}{An \code{secondp} object}

\item{...}{Other arguments}
}
\value{
Summary of a model
}
\description{
\code{summary.secondp}: Summary of the final model
}
